+++ src/renderer/containers/PreferencesDialog.tsx
@@ -1,5 +1,16 @@
 import * as React from 'react';
-import { AnchorButton, ControlGroup, Intent, Switch, Tab, Tabs } from '@blueprintjs/core';
+import {
+    AnchorButton,
+    Button,
+    Card,
+    Collapse,
+    ControlGroup,
+    HTMLTable,
+    Intent,
+    Switch,
+    Tab,
+    Tabs
+} from '@blueprintjs/core';
 import { connect, DispatchProp } from 'react-redux';
 
 import { SessionState, State, WebAPIServiceInfo } from '../state';
@@ -17,8 +28,8 @@ import { CATE_APP_VERSION } from '../../config';
 interface IPreferencesDialogProps {
     isOpen: boolean;
     preferences: SessionState;
-    serviceInfo: WebAPIServiceInfo,
-    serviceURL: string,
+    serviceInfo: WebAPIServiceInfo;
+    serviceURL: string;
 }
 
 function mapStateToProps(state: State): IPreferencesDialogProps {
@@ -31,10 +42,10 @@ function mapStateToProps(state: State): IPreferencesDialogProps {
 }
 
 
-const ABOUT_LABEL_STYLE = {width: '40%'};
-const ABOUT_CONTROL_GROUP_STYLE = {width: '100%', marginBottom: '1em'};
-const ABOUT_DIV_STYLE = {width: '100%', marginTop: '1em'};
-const ABOUT_ITEM_STYLE: React.CSSProperties = {width: '60%', wordWrap: 'break-word'};
+const SYSTEM_LABEL_STYLE = {width: '40%'};
+const SYSTEM_CONTROL_GROUP_STYLE = {width: '100%', marginBottom: '1em'};
+const SYSTEM_DIV_STYLE = {width: '100%', marginTop: '1em'};
+const SYSTEM_ITEM_STYLE: React.CSSProperties = {width: '60%', wordWrap: 'break-word'};
 
 class PreferencesDialog extends React.Component<IPreferencesDialogProps & DispatchProp<State>, SessionState> {
     static readonly DIALOG_ID = 'preferencesDialog';
@@ -92,6 +103,7 @@ class PreferencesDialog extends React.Component<IPreferencesDialogProps & Dispat
             <ModalDialog
                 isOpen={this.props.isOpen}
                 title={PreferencesDialog.DIALOG_TITLE}
+                style={{minWidth: '30em'}}
                 icon="confirm"
                 onCancel={this.onCancel}
                 onConfirm={this.onConfirm}
@@ -111,7 +123,7 @@ class PreferencesDialog extends React.Component<IPreferencesDialogProps & Dispat
                 <Tab id="g" title="General" panel={this.renderGeneralPanel()}/>
                 <Tab id="dm" title="Data Management" panel={this.renderDataManagementPanel()}/>
                 <Tab id="pc" title="Proxy Configuration" panel={this.renderProxyConfigurationPanel()}/>
-                <Tab id="a" title="About" panel={this.renderAboutPanel()}/>
+                <Tab id="a" title="System" panel={this.renderSystemPanel()}/>
             </Tabs>
         );
     }
@@ -148,24 +160,68 @@ class PreferencesDialog extends React.Component<IPreferencesDialogProps & Dispat
         );
     }
 
-    private renderAboutPanel() {
+    private renderSystemPanel() {
+        const showProperties = this.getStateValue('showProperties', false);
+        const onShowPropertiesChange = this.getChangeHandler('showProperties', false);
+
         return (
-            <div style={ABOUT_DIV_STYLE}>
-                <ControlGroup style={ABOUT_CONTROL_GROUP_STYLE}>
-                    <span style={ABOUT_LABEL_STYLE}>Cate UI version:</span>
-                    <span style={ABOUT_ITEM_STYLE}><code>{CATE_APP_VERSION}</code></span>
+            <div style={SYSTEM_DIV_STYLE}>
+                <ControlGroup style={SYSTEM_CONTROL_GROUP_STYLE}>
+                    <span style={SYSTEM_LABEL_STYLE}>Cate UI version:</span>
+                    <span style={SYSTEM_ITEM_STYLE}><code>{CATE_APP_VERSION}</code></span>
                 </ControlGroup>
-                <ControlGroup style={ABOUT_CONTROL_GROUP_STYLE}>
-                    <span style={ABOUT_LABEL_STYLE}>Cate API URL:</span>
-                    <span style={ABOUT_ITEM_STYLE}><code>{this.props.serviceURL}</code></span>
+                <ControlGroup style={SYSTEM_CONTROL_GROUP_STYLE}>
+                    <span style={SYSTEM_LABEL_STYLE}>Cate API URL:</span>
+                    <span style={SYSTEM_ITEM_STYLE}><code>{this.props.serviceURL}</code></span>
                 </ControlGroup>
-                <ControlGroup style={ABOUT_CONTROL_GROUP_STYLE}>
-                    <span style={ABOUT_LABEL_STYLE}>Cate API version:</span>
-                    <span style={ABOUT_ITEM_STYLE}><code>{this.props.serviceInfo.version}</code></span>
+                <ControlGroup style={SYSTEM_CONTROL_GROUP_STYLE}>
+                    <span style={SYSTEM_LABEL_STYLE}>Cate API version:</span>
+                    <span style={SYSTEM_ITEM_STYLE}><code>{this.props.serviceInfo.version}</code></span>
                 </ControlGroup>
-                <ControlGroup style={ABOUT_CONTROL_GROUP_STYLE}>
-                    <span style={ABOUT_LABEL_STYLE}>File access mode:</span>
-                    <span style={ABOUT_ITEM_STYLE}>
+                {this.props.serviceInfo.dependencies && (
+                    <ControlGroup style={SYSTEM_CONTROL_GROUP_STYLE}>
+                        <span style={SYSTEM_LABEL_STYLE}>Cate dependencies:</span>
+                        <div>
+                            <Button
+                                onClick={() => onShowPropertiesChange(!showProperties)}
+                                small={true}
+                                style={{marginBottom: 2}}
+                            >
+                                {"Click to " + (showProperties ? "collapse" : "expand")}
+                            </Button>
+                            <Collapse isOpen={showProperties}>
+                                <Card>
+                                    <HTMLTable condensed={true} small={true} bordered={true}>
+                                        <thead>
+                                        <tr>
+                                            <th>Package</th>
+                                            <th>Version</th>
+                                        </tr>
+                                        </thead>
+                                        <tbody>
+                                        {
+                                            Object.getOwnPropertyNames(this.props.serviceInfo.dependencies!).map(
+                                                packageName => {
+                                                    const packageVersion = this.props.serviceInfo.dependencies[packageName];
+                                                    return (
+                                                        <tr>
+                                                            <td>{packageName}</td>
+                                                            <td>{packageVersion}</td>
+                                                        </tr>
+                                                    );
+                                                }
+                                            )
+                                        }
+                                        </tbody>
+                                    </HTMLTable>
+                                </Card>
+                            </Collapse>
+                        </div>
+                    </ControlGroup>
+                )}
+                <ControlGroup style={SYSTEM_CONTROL_GROUP_STYLE}>
+                    <span style={SYSTEM_LABEL_STYLE}>File access mode:</span>
+                    <span style={SYSTEM_ITEM_STYLE}>
                         {this.props.serviceInfo.userRootMode ? 'sandboxed file system' : 'full access to file system'}
                     </span>
                 </ControlGroup>
@@ -281,13 +337,13 @@ class PreferencesDialog extends React.Component<IPreferencesDialogProps & Dispat
     }
 
     private renderStringValue(propertyName: string, isBackend: boolean, label: string) {
-        const initialValue = this.getStateValue(propertyName, isBackend);
+        const value = this.getStateValue(propertyName, isBackend);
         const onChange = this.getChangeHandler(propertyName, isBackend);
         return (
             <ControlGroup style={{width: '100%', marginBottom: '1em', display: 'flex', alignItems: 'center'}}>
                 <span style={{flexGrow: 0.8}}>{label}:</span>
                 <TextField style={{flexGrow: 0.2}}
-                           value={initialValue}
+                           value={value}
                            onChange={onChange}
                 />
             </ControlGroup>
@@ -295,11 +351,11 @@ class PreferencesDialog extends React.Component<IPreferencesDialogProps & Dispat
     }
 
     private renderBooleanValue(propertyName: string, isBackend: boolean, label: string) {
-        const initialValue = this.getStateValue(propertyName, isBackend);
+        const value = this.getStateValue(propertyName, isBackend);
         const onChange = this.getChangeHandler(propertyName, isBackend);
         return (
             <div style={{width: '100%', marginBottom: '1em'}}>
-                <Switch checked={initialValue}
+                <Switch checked={value}
                         label={label}
                         onChange={(event: any) => onChange(event.target.checked)}/>
             </div>